cleanup sqlnorm
