+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Server/SqlNormalizer.cs
@@ -2,9 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-//devnote: perf optimization: consider changing the calls to Array.Reverse to inline unsafe code
-
 using System;
+using System.Data.SqlTypes;
 using System.Diagnostics;
 using System.IO;
 using System.Reflection;
@@ -13,45 +12,48 @@ using System.Runtime.InteropServices;
 
 namespace Microsoft.Data.SqlClient.Server
 {
-
     // The class that holds the offset, field, and normalizer for
     // a particular field.
     internal sealed class FieldInfoEx : IComparable
     {
-        internal readonly int offset;
-        internal readonly FieldInfo fieldInfo;
-        internal readonly Normalizer normalizer;
+        private readonly int _offset;
 
         internal FieldInfoEx(FieldInfo fi, int offset, Normalizer normalizer)
         {
-            this.fieldInfo = fi;
-            this.offset = offset;
+            _offset = offset;
+            FieldInfo = fi;
             Debug.Assert(normalizer != null, "normalizer argument should not be null!");
-            this.normalizer = normalizer;
+            Normalizer = normalizer;
         }
+
+        internal FieldInfo FieldInfo { get; private set; }
+        internal Normalizer Normalizer { get; private set; }
 
         // Sort fields by field offsets.
         public int CompareTo(object other)
         {
-            FieldInfoEx otherF = other as FieldInfoEx;
-            if (otherF == null)
-                return -1;
-            return this.offset.CompareTo(otherF.offset);
+            if (other is FieldInfoEx otherFieldInfoEx)
+            {
+                return _offset.CompareTo(otherFieldInfoEx._offset);
+            }
+            return -1;           
         }
     }
 
     // The most complex normalizer, a udt normalizer
     internal sealed class BinaryOrderedUdtNormalizer : Normalizer
     {
-        internal readonly FieldInfoEx[] m_fieldsToNormalize;
-        private int m_size;
-        private byte[] m_PadBuffer;
-        internal readonly object NullInstance;
+        private readonly FieldInfoEx[] _fieldsToNormalize;
+        private int _size;
+        private readonly byte[] _padBuffer;
+        private readonly object _nullInstance;
         //a boolean that tells us if a udt is a "top-level" udt,
         //i.e. one that does not require a null byte header.
-        private bool m_isTopLevelUdt;
+        private readonly bool _isTopLevelUdt;
 
+#if NETFRAMEWORK
         [System.Security.Permissions.ReflectionPermission(System.Security.Permissions.SecurityAction.Assert, MemberAccess = true)]
+#endif
         private FieldInfo[] GetFields(Type t)
         {
             return t.GetFields(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
@@ -59,42 +61,42 @@ namespace Microsoft.Data.SqlClient.Server
 
         internal BinaryOrderedUdtNormalizer(Type t, bool isTopLevelUdt)
         {
-            this.m_skipNormalize = false;
-            if (this.m_skipNormalize)
+            _skipNormalize = false;
+            if (_skipNormalize)
             {
-                //if skipping normalization, dont write the null
-                //byte header for IsNull
-                this.m_isTopLevelUdt = true;
+                // if skipping normalization, dont write the null
+                // byte header for IsNull
+                _isTopLevelUdt = true;
             }
-            //top level udt logic is disabled until we decide
-            //what to do about nested udts
-            this.m_isTopLevelUdt = true;
+
+            // top level udt logic is disabled until we decide
+            // what to do about nested udts
+            _isTopLevelUdt = true;
             //      else
             //        this.m_isTopLevelUdt = isTopLevelUdt;
-            //get all the fields
 
             FieldInfo[] fields = GetFields(t);
 
-            m_fieldsToNormalize = new FieldInfoEx[fields.Length];
+            _fieldsToNormalize = new FieldInfoEx[fields.Length];
 
             int i = 0;
 
             foreach (FieldInfo fi in fields)
             {
                 int offset = Marshal.OffsetOf(fi.DeclaringType, fi.Name).ToInt32();
-                m_fieldsToNormalize[i++] = new FieldInfoEx(fi, offset, GetNormalizer(fi.FieldType));
+                _fieldsToNormalize[i++] = new FieldInfoEx(fi, offset, GetNormalizer(fi.FieldType));
             }
 
             //sort by offset
-            Array.Sort(m_fieldsToNormalize);
+            Array.Sort(_fieldsToNormalize);
             //if this is not a top-level udt, do setup for null values.
             //null values need to compare less than all other values,
             //so prefix a null byte indicator.
-            if (!this.m_isTopLevelUdt)
+            if (!_isTopLevelUdt)
             {
                 //get the null value for this type, special case for sql types, which
                 //have a null field
-                if (typeof(System.Data.SqlTypes.INullable).IsAssignableFrom(t))
+                if (typeof(INullable).IsAssignableFrom(t))
                 {
                     PropertyInfo pi = t.GetProperty("Null",
                     BindingFlags.Public | BindingFlags.Static);
@@ -102,39 +104,31 @@ namespace Microsoft.Data.SqlClient.Server
                     {
                         FieldInfo fi = t.GetField("Null", BindingFlags.Public | BindingFlags.Static);
                         if (fi == null || fi.FieldType != t)
+                        {
                             throw new Exception("could not find Null field/property in nullable type " + t);
+                        }
                         else
-                            this.NullInstance = fi.GetValue(null);
+                        {
+                            _nullInstance = fi.GetValue(null);
+                        }
                     }
                     else
                     {
-                        this.NullInstance = pi.GetValue(null, null);
+                        _nullInstance = pi.GetValue(null, null);
                     }
-                    //create the padding buffer
-                    this.m_PadBuffer = new byte[this.Size - 1];
+
+                    _padBuffer = new byte[Size - 1];
                 }
             }
         }
 
-        internal bool IsNullable
-        {
-            get
-            {
-                return this.NullInstance != null;
-            }
-        }
+        internal bool IsNullable => _nullInstance != null;
 
         // Normalize the top-level udt
-        internal void NormalizeTopObject(object udt, Stream s)
-        {
-            Normalize(null, udt, s);
-        }
+        internal void NormalizeTopObject(object udt, Stream s) => Normalize(null, udt, s);
 
         // Denormalize a top-level udt and return it
-        internal object DeNormalizeTopObject(Type t, Stream s)
-        {
-            return DeNormalizeInternal(t, s);
-        }
+        internal object DeNormalizeTopObject(Type t, Stream s) => DeNormalizeInternal(t, s);
 
         // Prevent inlining so that reflection calls are not moved to caller that may be in a different assembly that may have a different grant set.
         [MethodImpl(MethodImplOptions.NoInlining)]
@@ -142,29 +136,29 @@ namespace Microsoft.Data.SqlClient.Server
         {
             object result = null;
             //if nullable and not the top object, read the null marker
-            if (!this.m_isTopLevelUdt && typeof(System.Data.SqlTypes.INullable).IsAssignableFrom(t))
+            if (!_isTopLevelUdt && typeof(System.Data.SqlTypes.INullable).IsAssignableFrom(t))
             {
                 byte nullByte = (byte)s.ReadByte();
                 if (nullByte == 0)
                 {
-                    result = this.NullInstance;
-                    s.Read(m_PadBuffer, 0, m_PadBuffer.Length);
+                    result = _nullInstance;
+                    s.Read(_padBuffer, 0, _padBuffer.Length);
                     return result;
                 }
             }
             if (result == null)
-                result = Activator.CreateInstance(t);
-            foreach (FieldInfoEx myField in m_fieldsToNormalize)
             {
-                myField.normalizer.DeNormalize(myField.fieldInfo, result, s);
+                result = Activator.CreateInstance(t);
+            }
+            foreach (FieldInfoEx field in _fieldsToNormalize)
+            {
+                field.Normalizer.DeNormalize(field.FieldInfo, result, s);
             }
             return result;
         }
 
         internal override void Normalize(FieldInfo fi, object obj, Stream s)
         {
-            //      if (fi != null)
-            //        Console.WriteLine("normalizing " + fi.FieldType + " pos " + s.Position);
             object inner;
             if (fi == null)
             {
@@ -175,14 +169,13 @@ namespace Microsoft.Data.SqlClient.Server
                 inner = GetValue(fi, obj);
             }
 
-            //If nullable and not the top object, write a null indicator
-            System.Data.SqlTypes.INullable oNullable = inner as System.Data.SqlTypes.INullable;
-            if (oNullable != null && !this.m_isTopLevelUdt)
+            // If nullable and not the top object, write a null indicator
+            if (inner is INullable nullable && !_isTopLevelUdt)
             {
-                if (oNullable.IsNull)
+                if (nullable.IsNull)
                 {
                     s.WriteByte(0);
-                    s.Write(m_PadBuffer, 0, m_PadBuffer.Length);
+                    s.Write(_padBuffer, 0, _padBuffer.Length);
                     return;
                 }
                 else
@@ -191,58 +184,38 @@ namespace Microsoft.Data.SqlClient.Server
                 }
             }
 
-            foreach (FieldInfoEx myField in m_fieldsToNormalize)
+            foreach (FieldInfoEx field in _fieldsToNormalize)
             {
-                myField.normalizer.Normalize(myField.fieldInfo, inner, s);
+                field.Normalizer.Normalize(field.FieldInfo, inner, s);
             }
         }
 
-        internal override void DeNormalize(FieldInfo fi, object recvr, Stream s)
-        {
-            SetValue(fi, recvr, DeNormalizeInternal(fi.FieldType, s));
-        }
+        internal override void DeNormalize(FieldInfo fi, object recvr, Stream s) => SetValue(fi, recvr, DeNormalizeInternal(fi.FieldType, s));
 
         internal override int Size
         {
             get
             {
-                if (m_size != 0)
-                    return m_size;
-                if (this.IsNullable && !this.m_isTopLevelUdt)
-                    m_size = 1;
-                foreach (FieldInfoEx myField in m_fieldsToNormalize)
+                if (_size != 0)
                 {
-                    m_size += myField.normalizer.Size;
+                    return _size;
                 }
-                return m_size;
+                if (IsNullable && !_isTopLevelUdt)
+                {
+                    _size = 1;
+                }
+                foreach (FieldInfoEx field in _fieldsToNormalize)
+                {
+                    _size += field.Normalizer.Size;
+                }
+                return _size;
             }
         }
     }
 
     internal abstract class Normalizer
     {
-        /*
-        protected internal static string GetString(byte[] array)
-        {
-          StringBuilder sb = new StringBuilder();
-          //sb.Append("0x");
-          foreach (byte b in array)
-          {
-            sb.Append(b.ToString("X2", CultureInfo.InvariantCulture));
-          }
-          return sb.ToString();
-        }
-        */
-
-        protected bool m_skipNormalize;
-
-        /*
-        internal static bool IsByteOrderedUdt(Type t)
-        {
-          SqlUserDefinedTypeAttribute a = SerializationHelper.GetUdtAttribute(t);
-          return a.IsByteOrdered;
-        }
-        */
+        protected bool _skipNormalize;
 
         internal static Normalizer GetNormalizer(Type t)
         {
@@ -277,31 +250,32 @@ namespace Microsoft.Data.SqlClient.Server
                 n = new BinaryOrderedUdtNormalizer(t, false);
             }
             if (n == null)
-                throw new Exception(StringsHelper.GetString(Strings.Sql_CanotCreateNormalizer, t.FullName));
-            n.m_skipNormalize = false;
+            {
+                throw new Exception(StringsHelper.GetString(Strings.SQL_CannotCreateNormalizer, t.FullName));
+            }
+            n._skipNormalize = false;
             return n;
         }
 
         internal abstract void Normalize(FieldInfo fi, object recvr, Stream s);
+
         internal abstract void DeNormalize(FieldInfo fi, object recvr, Stream s);
 
         protected void FlipAllBits(byte[] b)
         {
             for (int i = 0; i < b.Length; i++)
+            {
                 b[i] = (byte)~b[i];
+            }
         }
-
+#if NETFRAMEWORK
         [System.Security.Permissions.ReflectionPermission(System.Security.Permissions.SecurityAction.Assert, MemberAccess = true)]
-        protected object GetValue(FieldInfo fi, object obj)
-        {
-            return fi.GetValue(obj);
-        }
-
+#endif
+        protected object GetValue(FieldInfo fi, object obj) => fi.GetValue(obj);
+#if NETFRAMEWORK
         [System.Security.Permissions.ReflectionPermission(System.Security.Permissions.SecurityAction.Assert, MemberAccess = true)]
-        protected void SetValue(FieldInfo fi, object recvr, object value)
-        {
-            fi.SetValue(recvr, value);
-        }
+#endif 
+        protected void SetValue(FieldInfo fi, object recvr, object value) => fi.SetValue(recvr, value);
 
         internal abstract int Size { get; }
     }
@@ -311,8 +285,6 @@ namespace Microsoft.Data.SqlClient.Server
         internal override void Normalize(FieldInfo fi, object obj, Stream s)
         {
             bool b = (bool)GetValue(fi, obj);
-            //      Console.WriteLine("normalized " + fi.FieldType + " " + fi.GetValue(obj)
-            //        + " to " + (b?"01":"00") + " pos " + s.Position);
             s.WriteByte((byte)(b ? 1 : 0));
         }
 
@@ -322,13 +294,9 @@ namespace Microsoft.Data.SqlClient.Server
             SetValue(fi, recvr, b == 1);
         }
 
-        internal override int Size { get { return 1; } }
+        internal override int Size => 1;
     }
 
-    // I could not find a simple way to convert a sbyte to a byte
-    // and vice versa in the framework api. Convert.ToSByte() checks that
-    // the value is in range.
-    // So, we just do the conversion inline.
     internal sealed class SByteNormalizer : Normalizer
     {
         internal override void Normalize(FieldInfo fi, object obj, Stream s)
@@ -339,16 +307,20 @@ namespace Microsoft.Data.SqlClient.Server
             {
                 b = (byte)sb;
             }
-            if (!this.m_skipNormalize)
+            if (!_skipNormalize)
+            {
                 b ^= 0x80; //flip the sign bit
+            }
             s.WriteByte(b);
         }
 
         internal override void DeNormalize(FieldInfo fi, object recvr, Stream s)
         {
             byte b = (byte)s.ReadByte();
-            if (!this.m_skipNormalize)
+            if (!_skipNormalize)
+            {
                 b ^= 0x80; //flip the sign bit
+            }
             sbyte sb;
             unchecked
             {
@@ -357,7 +329,7 @@ namespace Microsoft.Data.SqlClient.Server
             SetValue(fi, recvr, sb);
         }
 
-        internal override int Size { get { return 1; } }
+        internal override int Size => 1;
     }
 
     internal sealed class ByteNormalizer : Normalizer
@@ -374,7 +346,7 @@ namespace Microsoft.Data.SqlClient.Server
             SetValue(fi, recvr, b);
         }
 
-        internal override int Size { get { return 1; } }
+        internal override int Size => 1;
     }
 
     internal sealed class ShortNormalizer : Normalizer
@@ -382,7 +354,7 @@ namespace Microsoft.Data.SqlClient.Server
         internal override void Normalize(FieldInfo fi, object obj, Stream s)
         {
             byte[] b = BitConverter.GetBytes((short)GetValue(fi, obj));
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 Array.Reverse(b);
                 b[0] ^= 0x80;
@@ -392,9 +364,9 @@ namespace Microsoft.Data.SqlClient.Server
 
         internal override void DeNormalize(FieldInfo fi, object recvr, Stream s)
         {
-            byte[] b = new Byte[2];
+            byte[] b = new byte[2];
             s.Read(b, 0, b.Length);
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 b[0] ^= 0x80;
                 Array.Reverse(b);
@@ -410,7 +382,7 @@ namespace Microsoft.Data.SqlClient.Server
         internal override void Normalize(FieldInfo fi, object obj, Stream s)
         {
             byte[] b = BitConverter.GetBytes((ushort)GetValue(fi, obj));
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 Array.Reverse(b);
             }
@@ -419,16 +391,16 @@ namespace Microsoft.Data.SqlClient.Server
 
         internal override void DeNormalize(FieldInfo fi, object recvr, Stream s)
         {
-            byte[] b = new Byte[2];
+            byte[] b = new byte[2];
             s.Read(b, 0, b.Length);
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 Array.Reverse(b);
             }
             SetValue(fi, recvr, BitConverter.ToUInt16(b, 0));
         }
 
-        internal override int Size { get { return 2; } }
+        internal override int Size => 2;
     }
 
     internal sealed class IntNormalizer : Normalizer
@@ -436,21 +408,19 @@ namespace Microsoft.Data.SqlClient.Server
         internal override void Normalize(FieldInfo fi, object obj, Stream s)
         {
             byte[] b = BitConverter.GetBytes((int)GetValue(fi, obj));
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 Array.Reverse(b);
                 b[0] ^= 0x80;
             }
-            //      Console.WriteLine("normalized " + fi.FieldType + " " + fi.GetValue(obj)
-            //        + " to " + GetString(b) + " pos " + s.Position);
             s.Write(b, 0, b.Length);
         }
 
         internal override void DeNormalize(FieldInfo fi, object recvr, Stream s)
         {
-            byte[] b = new Byte[4];
+            byte[] b = new byte[4];
             s.Read(b, 0, b.Length);
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 b[0] ^= 0x80;
                 Array.Reverse(b);
@@ -458,7 +428,7 @@ namespace Microsoft.Data.SqlClient.Server
             SetValue(fi, recvr, BitConverter.ToInt32(b, 0));
         }
 
-        internal override int Size { get { return 4; } }
+        internal override int Size => 4;
     }
 
     internal sealed class UIntNormalizer : Normalizer
@@ -466,7 +436,7 @@ namespace Microsoft.Data.SqlClient.Server
         internal override void Normalize(FieldInfo fi, object obj, Stream s)
         {
             byte[] b = BitConverter.GetBytes((uint)GetValue(fi, obj));
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 Array.Reverse(b);
             }
@@ -477,14 +447,14 @@ namespace Microsoft.Data.SqlClient.Server
         {
             byte[] b = new byte[4];
             s.Read(b, 0, b.Length);
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 Array.Reverse(b);
             }
             SetValue(fi, recvr, BitConverter.ToUInt32(b, 0));
         }
 
-        internal override int Size { get { return 4; } }
+        internal override int Size => 4;
     }
 
     internal sealed class LongNormalizer : Normalizer
@@ -492,7 +462,7 @@ namespace Microsoft.Data.SqlClient.Server
         internal override void Normalize(FieldInfo fi, object obj, Stream s)
         {
             byte[] b = BitConverter.GetBytes((long)GetValue(fi, obj));
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 Array.Reverse(b);
                 b[0] ^= 0x80;
@@ -502,9 +472,9 @@ namespace Microsoft.Data.SqlClient.Server
 
         internal override void DeNormalize(FieldInfo fi, object recvr, Stream s)
         {
-            byte[] b = new Byte[8];
+            byte[] b = new byte[8];
             s.Read(b, 0, b.Length);
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 b[0] ^= 0x80;
                 Array.Reverse(b);
@@ -512,7 +482,7 @@ namespace Microsoft.Data.SqlClient.Server
             SetValue(fi, recvr, BitConverter.ToInt64(b, 0));
         }
 
-        internal override int Size { get { return 8; } }
+        internal override int Size => 8;
     }
 
     internal sealed class ULongNormalizer : Normalizer
@@ -520,27 +490,25 @@ namespace Microsoft.Data.SqlClient.Server
         internal override void Normalize(FieldInfo fi, object obj, Stream s)
         {
             byte[] b = BitConverter.GetBytes((ulong)GetValue(fi, obj));
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 Array.Reverse(b);
             }
-            //      Console.WriteLine("normalized " + fi.FieldType + " " + fi.GetValue(obj)
-            //        + " to " + GetString(b));
             s.Write(b, 0, b.Length);
         }
 
         internal override void DeNormalize(FieldInfo fi, object recvr, Stream s)
         {
-            byte[] b = new Byte[8];
+            byte[] b = new byte[8];
             s.Read(b, 0, b.Length);
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 Array.Reverse(b);
             }
             SetValue(fi, recvr, BitConverter.ToUInt64(b, 0));
         }
 
-        internal override int Size { get { return 8; } }
+        internal override int Size => 8;
     }
 
     internal sealed class FloatNormalizer : Normalizer
@@ -549,7 +517,7 @@ namespace Microsoft.Data.SqlClient.Server
         {
             float f = (float)GetValue(fi, obj);
             byte[] b = BitConverter.GetBytes(f);
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 Array.Reverse(b);
                 if ((b[0] & 0x80) == 0)
@@ -566,7 +534,9 @@ namespace Microsoft.Data.SqlClient.Server
                     // Treat it same as positive zero, so that
                     // the normalized key will compare equal.
                     if (f < 0)
+                    {
                         FlipAllBits(b);
+                    }
                 }
             }
             s.Write(b, 0, b.Length);
@@ -574,9 +544,9 @@ namespace Microsoft.Data.SqlClient.Server
 
         internal override void DeNormalize(FieldInfo fi, object recvr, Stream s)
         {
-            byte[] b = new Byte[4];
+            byte[] b = new byte[4];
             s.Read(b, 0, b.Length);
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 if ((b[0] & 0x80) > 0)
                 {
@@ -594,7 +564,7 @@ namespace Microsoft.Data.SqlClient.Server
             SetValue(fi, recvr, BitConverter.ToSingle(b, 0));
         }
 
-        internal override int Size { get { return 4; } }
+        internal override int Size => 4;
     }
 
     internal sealed class DoubleNormalizer : Normalizer
@@ -603,7 +573,7 @@ namespace Microsoft.Data.SqlClient.Server
         {
             double d = (double)GetValue(fi, obj);
             byte[] b = BitConverter.GetBytes(d);
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 Array.Reverse(b);
                 if ((b[0] & 0x80) == 0)
@@ -624,16 +594,14 @@ namespace Microsoft.Data.SqlClient.Server
                     }
                 }
             }
-            //      Console.WriteLine("normalized " + fi.FieldType + " " + fi.GetValue(obj)
-            //        + " to " + GetString(b));
             s.Write(b, 0, b.Length);
         }
 
         internal override void DeNormalize(FieldInfo fi, object recvr, Stream s)
         {
-            byte[] b = new Byte[8];
+            byte[] b = new byte[8];
             s.Read(b, 0, b.Length);
-            if (!m_skipNormalize)
+            if (!_skipNormalize)
             {
                 if ((b[0] & 0x80) > 0)
                 {
@@ -651,6 +619,6 @@ namespace Microsoft.Data.SqlClient.Server
             SetValue(fi, recvr, BitConverter.ToDouble(b, 0));
         }
 
-        internal override int Size { get { return 8; } }
+        internal override int Size => 8;
     }
 }
